requirements/base.txt
django==4.2.4
django-celery-beat==2.5.0
djangorestframework==3.14.0
django-redis==5.4.0
django-extensions==3.2.3
django-environ==0.11.2
django-cors-headers==4.3.0
django-filter==23.3
celery==5.3.4
flower==2.0.1
gunicorn==21.2.0
redis==4.6.0
requests==2.31.0
pytest==7.4.2
pytest-django==4.5.2
sentry-sdk==1.32.0
whitenoise==6.5.0
responses==0.23.1
flake8==3.7.9
black==19.10b0
factory-boy==3.3.0
coreapi==2.3.3
drf-yasg==1.21.7